normalize plot_phyto_size
normalized phytoplankton size class data to 1 to alleviate issues in data submission. kim's methods do not produce data that correctly sum to 1. to resolve this in the plotting routine, climatology data were normalized to a new scale such that they add up to 1 while maintaining their relative proportions to one another. phytoplankton size classes from 2023 were normalized to be on the same scale as the climatology.