2025 update to engagement
Note: the column `EJ "Rating` has not been included in this year's submission. If the bubble plot is something we plan to continue visualizing, we will need to ask SSB about this column. Bobby's community social vulnerability dataset contains these ratings for the smaller categories, so it may be derivable from that.